+++ naomi/profile.py
@@ -5,6 +5,8 @@ tell whether an option is configured or not, or the actual value
 """
 from naomi import paths
 import os
+from . import populate
+import re
 import yaml
 
 _profile = {}
@@ -31,6 +33,107 @@ def get_profile(command=""):
         raise ValueError("command '{}' not understood".format(command))
     if not _profile_read:
         # Open and read the profile
+        # Create .naomi dir if it does not exist yet
+        if not os.path.exists(paths.SUB_PATH):
+            try:
+                os.makedirs(paths.SUB_PATH)
+            except OSError:
+                self._logger.error("Could not create .naomi dir: '%s'",
+                                   paths.SUB_PATH, exc_info=True)
+                raise
+
+        # Check if .naomi dir is writable
+        if not os.access(paths.SUB_PATH, os.W_OK):
+            self._logger.critical(
+                " ".join([
+                    ".naomi dir {:s} is not writable. Naomi",
+                    "won't work correctly."
+                ]).format(
+                    paths.SUB_PATH
+                )
+            )
+        # Create .naomi/configs dir if it does not exist yet
+        if not os.path.exists(paths.CONFIG_PATH):
+            try:
+                os.makedirs(paths.CONFIG_PATH)
+            except OSError:
+                self._logger.error("Could not create .naomi/configs dir: '%s'",
+                                   paths.CONFIG_PATH, exc_info=True)
+                raise
+
+        # Check if .naomi/configs dir is writable
+        if not os.access(paths.CONFIG_PATH, os.W_OK):
+            self._logger.critical(
+                " ".join([
+                    ".naomi/configs dir {:s} is not writable. Naomi",
+                    "won't work correctly."
+                ]).format(
+                    paths.CONFIG_PATH
+                )
+            )
+        # For backwards compatibility, move old profile.yml to newly
+        # created config dir
+        old_configfile = paths.sub('profile.yml')
+        new_configfile = paths.sub(os.path.join('configs','profile.yml'))
+        if os.path.exists(old_configfile):
+            if os.path.exists(new_configfile):
+                self._logger.warning(
+                    " ".join([
+                        "Deprecated profile file found: '{:s}'. ",
+                        "Please remove it."
+                    ]).format(old_configfile)
+                )
+            else:
+                self._logger.warning(
+                    " ".join([
+                        "Deprecated profile file found: '{:s}'.",
+                        "Trying to move it to new location '{:s}'."
+                    ]).format(
+                        old_configfile,
+                        new_configfile
+                    )
+                )
+                try:
+                    shutil.move(old_configfile, new_configfile)
+                except shutil.Error:
+                    self._logger.error(
+                        " ".join([
+                            "Unable to move config file.",
+                            "Please move it manually.",
+                            "“{} → {}”".format(old_configfile,new_configfile)
+                        ]),
+                        exc_info=True
+                    )
+                    raise
+
+        # Read config
+        # set a loop so we can keep looping back until the config file exists
+        config_read = False
+        while(not config_read):
+            try:
+                with open(new_configfile, "r") as f:
+                    _profile = yaml.safe_load(f)
+                    _profile_read = True
+                    config_read = True
+            except IOError:
+                print("Your config file does not exist.")
+                text_input = input(
+                    " ".join([
+                        "Would you like to answer a few ",
+                        "questions to create a new one? "
+                    ])
+                )
+                if(re.match(r'\s*[Yy]', text_input)):
+                    _profile_read = True
+                    populate.run()
+                    config_read = True
+                else:
+                    print("Cannot continue. Exiting.")
+                    quit()
+            except (yaml.parser.ParserError, yaml.scanner.ScannerError) as e:
+                self._logger.error("Unable to parse config file: %s %s",
+                                e.problem.strip(), str(e.problem_mark).strip())
+                raise
         configfile = paths.config('profile.yml')
         with open(configfile, "r") as f:
             _profile = yaml.safe_load(f)
